Move initializeEmbeddding to initializeInteractiveEmbedding in embedding/interactive-embedding
