+++ e2e/support/helpers/e2e-embedding-iframe-interactive-helpers.ts
@@ -47,7 +47,7 @@ export function postMessageToIframe(options: PostMessageOptions) {
   const { messageData, iframeSelector } = options;
 
   // this madness is necessary to simulate a real MessageEvent coming from the parent window
-  // because in frontend/src/metabase/utils/embed.js we check e.source === window.parent
+  // because in frontend/src/metabase/embedding/interactive-embedding/initialize we check e.source === window.parent
   cy.get(iframeSelector)
     .should("be.visible")
     .then(($iframe) => {
