+++ frontend/src/metabase/app.js
@@ -33,13 +33,13 @@ import { initializePlugins } from "ee-plugins";
 import { AppThemeProvider } from "metabase/AppThemeProvider";
 import { ModifiedBackend } from "metabase/common/components/dnd/ModifiedBackend";
 import registerDashboardVisualizations from "metabase/dashboard/visualizations/register";
+import { initializeInteractiveEmbedding } from "metabase/embedding/interactive-embedding";
 import { MetabotProvider } from "metabase/metabot/context";
 import { PLUGIN_APP_INIT_FUNCTIONS } from "metabase/plugins";
 import { refreshSiteSettings } from "metabase/redux/settings";
 import { GlobalStyles } from "metabase/styled-components/containers/GlobalStyles";
 import { EmotionCacheProvider } from "metabase/ui/components/theme/EmotionCacheProvider";
 import api from "metabase/utils/api";
-import { initializeEmbedding } from "metabase/utils/embed";
 import { captureConsoleErrors } from "metabase/utils/errors";
 import { initTracing, rotateTraceId } from "metabase/utils/otel";
 import { MetabaseReduxProvider } from "metabase/utils/redux/custom-context";
@@ -79,7 +79,7 @@ function _init(reducers, getRoutes, callback) {
     syncedHistory.listen(() => rotateTraceId());
   }
 
-  initializeEmbedding(store);
+  initializeInteractiveEmbedding(store.dispatch);
 
   const root = createRoot(document.getElementById("root"));
