+++ frontend/src/metabase/embedding/interactive-embedding/index.ts
@@ -1,0 +1,1 @@
+export * from "./initialize";
